% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blade_params.R
\name{blade_params}
\alias{blade_params}
\title{Create a search grid}
\usage{
blade_params(params, ...)
}
\arguments{
\item{params}{A list of name-value pairs. The name will become the column name in the output.}

\item{...}{Conditions to filter the grid. Optional.}
}
\value{
A tibble with one column for each input in ...., optionally filtered by the conditions, and beginning with a column identifying each test case.
}
\description{
\code{blade_params()} takes a list of name-value pairs and sets up a grid comprising all possible combinations of these values. It also provides the optional ability to filter out certain combinations through specifying conditions.
}
\details{
This function provides a wrapper around tidyr::expand_grid and then (optionally) filters it by passing on the given conditions to dplyr::filter().
}
\examples{
params <- list(
  a = seq(1:3),
  b = 2
)
blade_params(params, a > b)
}
