% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blade_params.R
\name{blade_partition}
\alias{blade_partition}
\title{FUNCTION_TITLE}
\usage{
blade_partition(grid, from, to = NA)
}
\arguments{
\item{grid}{DESCRIPTION.}

\item{from}{DESCRIPTION.}

\item{to}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
