% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blade_runr.R
\name{blade_runr}
\alias{blade_runr}
\title{Execute your test functions smoothly with control}
\usage{
blade_runr(grid, use_sound = TRUE)
}
\arguments{
\item{grid}{A grid of test parameters, ideally generated by \code{blade_params()}.}

\item{use_sound}{logical. Whether sounds should be played at the completion of tests.}
}
\value{
The value of the post_runr function, as specified in \code{blade_setup()}
}
\description{
\code{blade_runr()} sets your tests in motion. After setting them up with \code{blade_setup()}, this function takes a grid of parameters generated by \code{blade_params()} and iterates over it, executing your functions for each individual test case.
}
\examples{
blade_setup(
  run_name = "test_run",
  runr = function() 1,
)
params <- list(
  a = seq(1:3),
  b = 2
)
grid <- blade_params(params)
blade_runr(grid)
}
