% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Create a search grid}
\usage{
create_grid(params, ...)
}
\arguments{
\item{params}{A list of name-value pairs. The name will become the column name in the output.}

\item{...}{Conditions to be passed on to dplyr::filter, if required.}
}
\value{
A tibble with one column for each input in ...., optionally filtered by the conditions
}
\description{
This function provides a wrapper around tidyr::expand_grid and filters it by the given filter conditions (optional).
}
\examples{
params <- list(
  a = seq(1:3),
  b = 2
)
create_grid(params, a > b)
}
