% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config}
\alias{get_config}
\title{Set Config}
\usage{
get_config(attr)
}
\arguments{
\item{attr}{Name of attribute to get/set}
}
\description{
\code{Set_config} manages the internal configuration for bladerunr.
}
