% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{getset}
\alias{getset}
\alias{set_config}
\alias{get_config}
\title{Get and Set Config}
\usage{
set_config(attr, value)

get_config(attr)
}
\arguments{
\item{attr}{Name of attribute to get/set}

\item{value}{Value for named attribute}
}
\value{
List of config settings.
}
\description{
These two functions manage the internal configuration for bladerunr.
They are not exposed to the user.
}
