% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_config}
\alias{set_config}
\title{Get Config}
\usage{
set_config(...)
}
\arguments{
\item{...}{name = value pairs}
}
\value{
List of config settings.
}
\description{
\code{get_config} manages the internal configuration for bladerunr.
}
