% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_calls.R
\name{setup_calls}
\alias{setup_calls}
\title{Specify callback functions to be executed before each model run}
\usage{
setup_calls(params, callbacks)
}
\arguments{
\item{params}{A list of params which will include all the parameters of the grid for the given test run.}

\item{callbacks}{A list of callback functions. Each takes one formal argument: params}
}
\description{
\code{setup_calls()} executes setup functions using the test parameters for a given run. This is where you specify how to use your parameters to update your model.
}
