% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blade_runr.R
\name{test_runr}
\alias{test_runr}
\title{Test your Runr}
\usage{
test_runr()
}
\description{
\code{test_runr()} enables you to test your runr setup before committing to a full run.
Bladerunr will take the first row of your grid and call your runr with the full
params and context that would be expected for each iteration.
}
