% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blade_setup.R
\name{blade_setup}
\alias{blade_setup}
\title{Setup bladerunr}
\usage{
blade_setup(
  run_name,
  runr,
  output_dir = NULL,
  timeout = NULL,
  max_attempts = 2,
  log_output = FALSE
)
}
\arguments{
\item{run_name}{The name of the run. Used for output folder naming.}

\item{runr}{A function to be executed as the main test. Must receive two arguments:
params and context. See vignette for more details.}

\item{output_dir}{A path to be given to the runr, if required. Optional.}

\item{timeout}{An int representing the time in seconds to allow the runr.
If the runr exceeds it, the function will either try again or move on depending on the \code{max_attempts} value. Optional.}

\item{max_attempts}{An int specifying the number of time to attempt a run
before moving on. Optional.}

\item{log_output}{If \code{TRUE}, runr output will be saved to a log in either your
\code{output_dir} or working directory if no \code{output_dir} was specified. Defaults to FALSE.}
}
\description{
\code{blade_setup()} is your first step. It sets up the function you want to
repeat for each iteration, as well as various other settings for your run.
}
\details{
Detailed explanation of how the runr and setup work can be found at the README.
}
\examples{

foo_run <- function() {
  # Do some fancy stuff
}

blade_setup("test-run", foo_run, "path/to/save/outputs")
}
