% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blade_setup.R
\name{blade_setup}
\alias{blade_setup}
\title{Setup bladerunr}
\usage{
blade_setup(
  run_name,
  runr,
  pre_runr = NULL,
  post_runr = NULL,
  output_dir = NULL,
  timeout = NULL,
  max_attempts = 2
)
}
\arguments{
\item{run_name}{The name of the run. Used for output folder naming.}

\item{runr}{A function to be executed as the main test. Function may produce side-effects, or return a value. Any value returned is passed to the post_runr. The function will also receive the following three arguments for context: test_number, run_name, and output_dir (as specified in the setup).}

\item{pre_runr}{A function to execute \emph{before} each model run. This function will each receive a list of all the model params for the given run as input. This function will be executed at every iteration. Optional.}

\item{post_runr}{A function to execute \emph{after} each model run. If the main runr returned a value, it will be passed on. The function will be executed at every iteration; and, if the user specifies an \code{output_dir}, this value will be passed to to the function as well. Optional.}

\item{output_dir}{A path to be given to the \code{post_runrs}, if required. Optional.}

\item{timeout}{An int representing the time in seconds to allow the runr. If the runr exceeds it, the function will either try again or move on depending on the \code{max_attempts} value. Optional.}

\item{max_attempts}{An int specifying the number of time to attempt a run before moving on. Optional.}
}
\description{
\code{blade_setup()} must be called to specify the functions
you want to repeat for each iteration of your grid search.
}
\details{
There are three types of callbacks bladerunr can use at each iteration:
pre-runrs, runrs, and post-runrs. This enables the user to break up their
functions into smaller chunks. It also anticipates that the model may involve
files being generated, either as inputs, outputs, or both; the callbacks
enable the user to deal with these as they wish.

Detailed explanation of how the runrs and setup works can be found at the vignette.
\code{vignette("help", package = "bladerunr")}
}
\examples{

foo_before <- function(params) {
  # Do some setup work
}
foo_run <- function() {
  # Do some fancy stuff
}
foo_after <- function() {
  # Transform and save model results
}

blade_setup("test-run", foo_before, foo_run, foo_after, "path/to/save/outputs")
}
