% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_config}
\alias{set_config}
\title{Get Config}
\usage{
set_config(attr, value)
}
\arguments{
\item{attr}{Name of attribute to get/set}

\item{value}{Value for named attribute}
}
\value{
List of config settings.
}
\description{
\code{get_config} manages the internal configuration for bladerunr.
}
